+++ docs/changelog.md
@@ -2,7 +2,11 @@ Sir. Clive Sinclairbot Changelog
 
 *WARNING* - releases marked with 🧠 were primarily written by the discord bot itself, via prompting its own large language model, so one is advised to treat accordingly.
 
-v0.0.5 🤯 Now uses `pnpm`. Added storage API for all plugins
+v0.0.7 ❕ LLM Chatbot plugin can now reset context, and help users
+
+v0.0.6 🤖 LLM Chatbot plugin, but still on unstable plugin API
+
+v0.0.5 🧱 Now uses `pnpm`. Added storage API for all plugins
 
 v0.0.4 🤯 Publicly usable, with example config, more plugins & some docs!